requirements.txt
tensorflow>=2.19
keras>=3.0
numpy
pandas
pillow
matplotlib
scikit-learn
rasterio
pyunpack
requests
jupyter
ipykernel